+++ core/src/test/java/nl/nn/adapterframework/pipes/ForEachChildElementPipeTest.java
@@ -1,8 +1,18 @@
 package nl.nn.adapterframework.pipes;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+
+import java.io.ByteArrayInputStream;
+import java.io.FilterInputStream;
+import java.io.IOException;
+import java.io.InputStream;
 
 import org.junit.Test;
+import org.xml.sax.Attributes;
+import org.xml.sax.ContentHandler;
+import org.xml.sax.SAXException;
+import org.xml.sax.helpers.XMLFilterImpl;
 
 import nl.nn.adapterframework.configuration.ConfigurationException;
 import nl.nn.adapterframework.core.IPipeLineSession;
@@ -11,6 +21,8 @@ import nl.nn.adapterframework.core.PipeLineSessionBase;
 import nl.nn.adapterframework.core.PipeRunException;
 import nl.nn.adapterframework.core.PipeRunResult;
 import nl.nn.adapterframework.core.PipeStartException;
+import nl.nn.adapterframework.core.SenderException;
+import nl.nn.adapterframework.parameters.ParameterResolutionContext;
 import nl.nn.adapterframework.senders.XsltSender;
 
 public class ForEachChildElementPipeTest extends PipeTestBase<ForEachChildElementPipe> {
@@ -31,8 +43,16 @@ public class ForEachChildElementPipeTest extends PipeTestBase<ForEachChildElemen
         return new ForEachChildElementPipe();
     }
 
-    protected ISender getElementRenderer() {
-    	XsltSender sender = new XsltSender();
+    protected ISender getElementRenderer(final SwitchCounter sc) {
+    	XsltSender sender = new XsltSender() {
+
+			@Override
+			public String sendMessage(String correlationID, String message, ParameterResolutionContext prc) throws SenderException {
+				if (sc!=null) sc.mark("out");
+				return super.sendMessage(correlationID, message, prc);
+			}
+    		
+    	};
     	sender.setXpathExpression("sub");
     	return sender;
     }
@@ -40,7 +60,7 @@ public class ForEachChildElementPipeTest extends PipeTestBase<ForEachChildElemen
     
     @Test
     public void testBasic() throws PipeRunException, ConfigurationException, PipeStartException {
-    	pipe.setSender(getElementRenderer());
+    	pipe.setSender(getElementRenderer(null));
     	configurePipe();
     	pipe.start();
 
@@ -52,30 +72,146 @@ public class ForEachChildElementPipeTest extends PipeTestBase<ForEachChildElemen
 
     @Test
     public void testXPath() throws PipeRunException, ConfigurationException, PipeStartException {
+    	SwitchCounter sc = new SwitchCounter();
     	pipe.setElementXPathExpression("/root/sub");
     	pipe.setNamespaceAware(true);
-    	pipe.setSender(getElementRenderer());
+    	pipe.setSender(getElementRenderer(sc));
     	configurePipe();
     	pipe.start();
 
-        PipeRunResult prr = pipe.doPipe(messageBasic, session);
+		ByteArrayInputStream bais = new ByteArrayInputStream(messageBasic.getBytes());    	
+        PipeRunResult prr = pipe.doPipe(new LoggingInputStream(bais,sc), session);
         String actual=prr.getResult().toString();
         
         assertEquals(expectedBasic, actual);
+		assertTrue("streaming failure: switch count ["+sc.count+"] should be larger than 2",sc.count>2);
     }
 
     @Test
     public void testXPathWithSpecialChars() throws PipeRunException, ConfigurationException, PipeStartException {
+    	SwitchCounter sc = new SwitchCounter();
     	pipe.setElementXPathExpression("/root/sub[position()<3]");
     	pipe.setNamespaceAware(true);
-    	pipe.setSender(getElementRenderer());
+    	pipe.setSender(getElementRenderer(sc));
     	configurePipe();
     	pipe.start();
 
-        PipeRunResult prr = pipe.doPipe(messageBasic, session);
+		ByteArrayInputStream bais = new ByteArrayInputStream(messageBasic.getBytes());    	
+        PipeRunResult prr = pipe.doPipe(new LoggingInputStream(bais,sc), session);
         String actual=prr.getResult().toString();
         
         assertEquals(expectedBasic, actual);
+		assertTrue("streaming failure: switch count ["+sc.count+"] should be larger than 2",sc.count>2);
     }
+    
+    
+	private class SwitchCounter {
+		public int count;
+		private String prevLabel;
+		
+		public void mark(String label) {
+			if (prevLabel==null || !prevLabel.equals(label)) {
+				prevLabel=label;
+				count++;
+			}
+		}
+	}
+
+	private class SaxLogger extends XMLFilterImpl implements ContentHandler {
+		
+		private String prefix;
+		private SwitchCounter sc;
+		
+		SaxLogger(String prefix, SwitchCounter sc) {
+			this.prefix=prefix;
+			this.sc=sc;
+		}
+		private void print(String string) {
+			log.debug(prefix+" "+string);
+			sc.mark(prefix);
+		}
+		
+		@Override
+		public void characters(char[] ch, int start, int length) throws SAXException {
+			print(new String(ch,start,length));
+			super.characters(ch, start, length);
+		}
+
+		@Override
+		public void startDocument() throws SAXException {
+			print("startDocument");
+			super.startDocument();
+		}
+
+		@Override
+		public void endDocument() throws SAXException {
+			print("endDocument");
+			super.endDocument();
+		}
+
+		@Override
+		public void startElement(String uri, String localName, String qName, Attributes attributes) throws SAXException {
+			print("startElement "+localName);
+			super.startElement(uri, localName, qName, attributes);
+		}
+		
+		@Override
+		public void endElement(String uri, String localName, String qName) throws SAXException {
+			print("endElement "+localName);
+			super.endElement(uri, localName, qName);
+		}
+
+		
+	}
+
+	private class LoggingInputStream extends FilterInputStream {
+
+		private int blocksize=10;
+		private SwitchCounter sc;
+		
+		public LoggingInputStream(InputStream arg0, SwitchCounter sc) {
+			super(arg0);
+			this.sc=sc;
+		}
+
+		private void print(String string) {
+			log.debug("in-> "+string);
+			sc.mark("in");
+		}
+
+		@Override
+		public int read() throws IOException {
+			int c=super.read();
+			print("in-> ["+((char)c)+"]");
+			return c;
+		}
+
+		@Override
+		public int read(byte[] buf, int off, int len) throws IOException {
+			int l=super.read(buf, off, len<blocksize?len:blocksize);
+			if (l<0) {
+				print("{EOF}");
+			} else {
+				print(new String(buf,off,l));
+			}
+			return l;
+		}
+
+		@Override
+		public int read(byte[] buf) throws IOException {
+			if (buf.length>blocksize) {
+				return read(buf,0,blocksize);
+			}
+			int l=super.read(buf);
+			if (l<0) {
+				print("{EOF}");
+			} else {
+				print(new String(buf,0,l));
+			}
+			return l;
+		}
+		
+	}
+
 
 }